chore: update vip numbering
