+++ simulations/vip-194/vip-194-testnet/abi/BEACON_ABI.json
